+++ SystemInventorys/SystemInventory-pojo/src/main/java/zhongchiedu/inventory/pojo/ProjectStockStatistics.java
@@ -31,16 +31,16 @@ public class ProjectStockStatistics extends GeneralBean<ProjectStockStatistics> 
 	private ProjectStock projectStock; //绑定库存商品
 	private String storageTime;//入库时间
 	private String depotTime;//出库时间
-	private long num;//出库、入库数量
-	private long newNum;//当前库存
-	private long actualPurchaseQuantity;//实际采购量 
+	private Integer num;//出库、入库数量
+	private Integer newNum;//当前库存
+	private Integer actualPurchaseQuantity;//实际采购量 
 	@DBRef
 	private Companys company;//绑定企业
 	@DBRef
 	private User user;//操作人
 	private boolean revoke; //是否可撤销
 	private boolean inOrOut;//入库还是出库
-	private long revokeNum;//撤销数量，默认为全部，撤销数量不能大于入库或者出库数量
+	private Integer revokeNum;//撤销数量，默认为全部，撤销数量不能大于入库或者出库数量
 	
 	private String personInCharge;//责任人
 	private String projectName;   //项目名
